requirements.txt
aioredis==1.2.0
amqp==2.4.1
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.2.1
Automat==0.7.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==2.2.5
Django==2.1.7
django-rest-swagger==2.2.0
djangorestframework==3.9.1
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
kombu==4.3.0
MarkupSafe==1.1.0
msgpack==0.6.1
openapi-codec==1.3.2
pkg-resources==0.0.0
PyHamcrest==1.9.0
pytz==2018.9
redis==3.2.0
requests==2.21.0
simplejson==3.16.0
six==1.12.0
Twisted==18.9.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.2.0
zope.interface==4.6.0
